cleaning up code error handling working on invite/update
to-do:
check for existing acct
clear search & auto filter w/o button
error handling
invites on event input
event search / single date logic

+++ src/app/event.service.ts
@@ -5,31 +5,47 @@ import {HttpService} from "./http.service";
 import {AccountService} from "./account.service";
 import {v4 as uuidv4} from "uuid";
 import {IAccount} from "./interfaces/IAccount";
+import {IInvite} from "./interfaces/IInvite";
 
 @Injectable({
   providedIn: 'root'
 })
 export class EventService {
+
+  //variables
   curEventList: IEvent[] =[];
   $eventList = new BehaviorSubject<IEvent[]>([]);
-  $event = new BehaviorSubject<IEvent | null>(null);
   $selectedEvent = new BehaviorSubject<IEvent | null>(null);
 
-  $eventError = new BehaviorSubject<string | null>(null);
-  private readonly EVENT_INVALID_EVENT_NAME = "You must provide a valid event name";
-  private readonly EVENT_HTTP_ERROR = "There was an error with the HTTP server";
   private account!: IAccount;
   private accountID: string = "";
 
+  $isEditing = new BehaviorSubject<boolean>(false);
+
+  $eventError = new BehaviorSubject<string | null>(null);
+
+  //error messages
+  private readonly EVENT_INVALID_EVENT_NAME = "You must provide a valid event name";
+  private readonly EVENT_HTTP_ERROR = "There was an error with the HTTP server";
+  private readonly EVENT_MISSING_VALUE = "There is a missing value";
+
+
   constructor(private httpService: HttpService,
               private accountService: AccountService) {
+
+    //get event list from server
     this.httpService.getEvents().pipe(first()).subscribe({
       next: eventList => {
         this.curEventList = eventList;
         this.$eventList.next(eventList)
+      },
+      error: (err) => {
+        this.$eventError.next(this.EVENT_HTTP_ERROR);
+        console.log(err)
       }
     })
 
+    //get account from account service
     this.accountService.$account.subscribe(account => {
       if (account) {
         this.account = account
@@ -37,39 +53,46 @@ export class EventService {
     })
   }
 
+  //delete selected event
   deleteEvent(event: IEvent){
-    // this.curEventList = this.curEventList.filter(events => events !== event)
-    // console.log('es delete works')
-    // console.log(this.curEventList);
-    // this.$eventList.next(this.curEventList);
-    this.httpService.deleteEvent(event.id).pipe(first()).subscribe({
-      next: (data) => {
-        console.log(data);
-        this.getEvents();
-        this.$selectedEvent.next(null);
-      },
-      error: (err) => {
-        console.error(err);
-      }
-    });
+    if (event.id == null){
+      this.$eventError.next(this.EVENT_MISSING_VALUE);
+    }
+    else{
+      this.httpService.deleteEvent(event.id).pipe(first()).subscribe({
+        next: (data) => {
+          console.log(data);
+          this.getEvents();
+          this.$selectedEvent.next(null);
+        },
+        error: (err) => {
+          this.$eventError.next(this.EVENT_HTTP_ERROR);
+          console.log(err)
+        }
+      });
+    }
   }
 
+  //get event list
   getEvents(){
     this.httpService.getEvents().pipe(first()).subscribe({
       next: eventList => {
         this.curEventList = eventList;
         this.$eventList.next(eventList)
+      },
+      error: (err) => {
+        this.$eventError.next(this.EVENT_HTTP_ERROR);
+        console.log(err)
       }
     })
   }
 
-  createEvent(eventForm: IEvent, dateConvert: Date, eventList: IAccount[]){
-    console.log('create event e s')
+  //create new event
+  createEvent(eventForm: IEvent, dateConvert: Date, eventList: IInvite[]){
     if (eventForm.eventName.length == 0){
       this.$eventError.next(this.EVENT_INVALID_EVENT_NAME)
     }
 
-    // @ts-ignore
     if (this.account.id != null){
       this.accountID = this.account?.id;
     }
@@ -81,23 +104,61 @@ export class EventService {
           eventName: eventForm.eventName,
           invited: {
             id: eventList}
-
         }
 
-    console.log(eventForm.eventDate)
+    //send event to http/server and listen for ob
     this.httpService.createEvent(event).pipe(first()).subscribe({
       next: (event) => {
-        //this.$event.next(event);
+        //TODO FIX?
+       //this.$event.next(event);
+        //get updated event list
         this.getEvents();
-        console.log('http adding ' + event);
       },
       error: (err) => {
         console.error(err)
         this.$eventError.next(this.EVENT_HTTP_ERROR)
       }
     })
-      }
+  }
 
+  //update from list clicked, change boolean & ensure selected event current
+  updateClick(event: IEvent){
+    this.$selectedEvent.next(event);
+    this.$isEditing.next(true);
+  }
+
+
+  //update selected event
+  updateEvent(editEvent: IEvent, dateConvert: Date, eventList: IInvite[]){
+    console.log(editEvent)
+
+    const event: IEvent  = {
+      id: "test",
+      creatorID: editEvent.creatorID,
+      eventDate: dateConvert,
+      eventName: editEvent.eventName,
+      invited: {
+        id: []}
+
+    }
+
+    this.httpService.updateEvent(event).pipe(first()).subscribe({
+      next: (event) => {
+        //TODO FIX?
+        //this.$event.next(event);
+        this.getEvents();
+        console.log('http updating ' + event);
+      },
+      error: (err) => {
+        console.error(err)
+        this.$eventError.next(this.EVENT_HTTP_ERROR)
+      }
+    })
+
+    this.$isEditing.next(false);
+  }
+
+  //search dates
   dateSearch(convertFromDate: Date, convertToDate: Date){
         console.log('e s datesearch' + convertFromDate + " " + convertToDate)
         this.$eventList.next(
@@ -109,12 +170,11 @@ export class EventService {
         ));
   }
 
+  //click on specific event from list and open details
   openEvent(event: IEvent) {
-    console.log("e s open event");
-    console.log(event)
     this.$selectedEvent.next(event);
-
   }
+
   closeEvent(){
     this.$selectedEvent.next(null);
   }
